# api/openapi.yaml
openapi: 3.1.2
info:
  title: Descendence API
  description: Basic api provided by the descendence api server
  version: v0.0.1
servers:
- url: http://127.0.0.1:8080
  description: Local development server for testing
paths:
...
  /api/v1/runs:
    post:
      operationId: createRun
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCreate"
      responses:
        "202":
          description: >
            Run accepted and queued. Never blocks on the container finishing -
            poll GET /api/v1/runs/{id} for the outcome. A repeated
            Idempotency-Key for the same principal returns the original run
            rather than creating a second one.
          headers:
            Location:
              description: "/api/v1/runs/{id} of the created (or original, if replayed) run"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Validation failed (e.g. empty argv, non-positive timeout)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    get:
      operationId: listRuns
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's nextCursor. Omit for the first page.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum runs to return.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: >
            Runs in descending queued_at order (newest first), keyset-paginated -
            never offset-based, since the table grows forever and offset
            pagination skips rows under concurrent inserts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
        "400":
          description: Malformed cursor, or limit outside 1-200
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: >
        Scoped per principal. A retried request with the same key returns the
        original run instead of creating a second one.
      schema:
        type: string
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      required:
        - type
        - title
        - status
    RunCreate:
      type: object
      properties:
        imageRef:
          type: string
          description: OCI image reference, e.g. "docker.io/library/alpine:latest"
        argv:
          type: array
          items:
            type: string
          minItems: 1
          description: >
            Executed as-is, never interpolated into a shell string - a value
            like "; rm -rf /" is passed through literally and harmlessly.
        timeoutSeconds:
          type: integer
          minimum: 1
          default: 3600
      required:
        - imageRef
        - argv
    Run:
      type: object
      properties:
        id:
          type: integer
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled, lost]
        imageRef:
          type: string
        argv:
          type: array
          items:
            type: string
        timeoutSeconds:
          type: integer
        containerId:
          type: [string, "null"]
        exitCode:
          type: [integer, "null"]
        failureReason:
          type: [string, "null"]
        cancelRequestedAt:
          description: >
            When cancellation was requested. Set before the run is actually
            cancelled - a running run's container still has to be stopped by
            the supervisor - so a run can be `running` with this set, meaning
            "on its way out". Also kept on a run that finished on its own
            before the request landed, which is the case worth being able to
            explain afterwards.
          type: [string, "null"]
          format: date-time
        queuedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        jobId:
          description: >
            The job this run came from, or null for an ad-hoc run created by
            POST /api/v1/runs. A job may since have been removed from its
            repository; the run still points at it, so that what ran remains
            explainable.
          type: [integer, "null"]
        commitSha:
          description: >
            The exact commit the job's manifest and script were read from,
            resolved when the run was created and never re-resolved. Together
            with jobId this is what makes a past run reproducible: check this
            commit out and you have precisely what executed. Null for an
            ad-hoc run.
          type: [string, "null"]
        runtimeId:
          description: >
            Set when the job named a runtime rather than an image directly
            (task 4.6). Resolved once, at creation, from the runtime's
            current image digest - rebuilding the runtime afterwards never
            changes this run's imageDigest.
          type: [integer, "null"]
        imageDigest:
          description: The runtime's image digest at the moment this run was created. Null for an ad-hoc run or a job that names an image directly.
          type: [string, "null"]
      required:
        - id
        - state
        - imageRef
        - argv
        - timeoutSeconds
        - queuedAt
    RunList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        nextCursor:
          description: Opaque cursor for the next page. Absent (or null) on the last page.
          type: [string, "null"]
      required:
        - items
...